requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
openai>=1.40.0
instructor>=1.5.0
pypdf>=4.0.0
python-multipart>=0.0.9
python-docx>=1.1.0
httpx>=0.27.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
crawl4ai>=0.4.0
feedparser>=6.0.0